post-build.sh should be idempotent, now that Docker-based builds can cache artifacts.
The procedure `append_inittab` needs it especially. Otherwise, `patch` aborts with "Reversed (or previously applied) patch detected".

+++ board/sharp/brain_imx28/post-build.sh
@@ -2,6 +2,9 @@
 
 append_inittab()
 {
+	if grep -q '^tty1::respawn:' "${TARGET_DIR}/etc/inittab"; then
+		return 0
+	fi
 	# --forward: don't reverse-apply
 	# -r -: discard rejected hunk
 	# -p 0: respect the entire path
@@ -25,7 +28,9 @@ promote_haveged_initscript()
 {
 	# haveged initializes at S21 by default. Move it to S10 so it starts before
 	# S20seedrng, which blocks on getrandom(2) until the CRNG is seeded.
-	mv "${TARGET_DIR}/etc/init.d/S21haveged" "${TARGET_DIR}/etc/init.d/S10haveged"
+	if [ -e "${TARGET_DIR}/etc/init.d/S21haveged" ]; then
+		mv "${TARGET_DIR}/etc/init.d/S21haveged" "${TARGET_DIR}/etc/init.d/S10haveged"
+	fi
 }
 
 main()
